Fixed some gcc warnings
<==== Mercurial data ====>
Branch: master

+++ client/src/device_plugins/archive/rtsp/rtsp_client_archive_delegate.cpp
@@ -336,7 +336,7 @@ QnAbstractMediaDataPtr QnRtspClientArchiveDelegate::getNextData()
     qint64 timeMs = AV_NOPTS_VALUE;
 	if (result && result->timestamp >= 0)
 		timeMs = result->timestamp/1000; // do not switch server if AV_NOPTS_VALUE and any other invalid packet timings
-    bool outOfRange = timeMs != AV_NOPTS_VALUE && ((m_rtspSession.getScale() >= 0 && timeMs >= m_serverTimePeriod.endTimeMs()) ||
+    bool outOfRange = (quint64)timeMs != AV_NOPTS_VALUE && ((m_rtspSession.getScale() >= 0 && timeMs >= m_serverTimePeriod.endTimeMs()) ||
                       (m_rtspSession.getScale() <  0 && timeMs < m_serverTimePeriod.startTimeMs));
     if (result == 0 || outOfRange || result->dataType == QnAbstractMediaData::EMPTY_DATA)
     {
